requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
loguru==0.7.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# APIs
keepa==1.3.15
tweepy==4.14.0

# Data processing
pandas==2.1.3
numpy==1.25.2

# Monitoring and health checks
psutil==5.9.6
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# CLI and display
tabulate==0.9.0
click==8.1.7

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1